Switch to ResourceNotFound exception

+++ src/main/java/xyz/bobby/unispring/repository/UserRepository.java
@@ -5,4 +5,4 @@ import org.springframework.stereotype.Repository;
 import xyz.bobby.unispring.model.User;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Long> {}
+public interface UserRepository extends JpaRepository<User, Integer> {}
